+++ CHANGELOG.md
@@ -152,6 +152,7 @@
 
 ### Mimirtool
 
+* [ENHANCEMENT] Add template render command to render locally a template. #7325
 * [ENHANCEMENT] Add `--extra-headers` option to `mimirtool rules` command to add extra headers to requests for auth. #7141
 * [ENHANCEMENT] Analyze Prometheus: set tenant header. #6737
 * [ENHANCEMENT] Add argument `--output-dir` to `mimirtool alertmanager get` where the config and templates will be written to and can be loaded via `mimirtool alertmanager load` #6760